+++ release.config.js
@@ -1,0 +1,10 @@
+const { releasercConfig } = require('@fairdataihub/config');
+
+const config = releasercConfig(
+  'default',
+  'fairdataihub',
+  'fairdataihub.org',
+  false,
+);
+
+module.exports = config;