+++ src/pages/_app.tsx
@@ -1,7 +1,9 @@
-import { AppProps } from 'next/app';
-import '@/styles/global.css';
+import { AppProps } from "next/app";
 
-import Layout from '../components/layout/layout';
+import "@/styles/global.css";
+import "@/styles/highlightjs.css";
+
+import Layout from "../components/layout/layout";
 
 export default function MyApp({ Component, pageProps }: AppProps) {
   return (