+++ src/pages/knowmore.tsx
@@ -1,13 +1,13 @@
-import Head from 'next/head';
+import Head from "next/head";
 
-import Hero from '@/components/knowmore/hero';
-import About from '@/components/knowmore/about';
-import Info from '@/components/knowmore/info';
-import Publications from '@/components/knowmore/publications';
+import Hero from "@/components/knowmore/hero";
+import About from "@/components/knowmore/about";
+import Info from "@/components/knowmore/info";
+import Publications from "@/components/knowmore/publications";
 
 export default function KnowMore() {
   return (
-    <div>
+    <>
       <Head>
         <title>KnowMore - Fair Data Innovations Hub</title>
         <meta
@@ -49,7 +49,7 @@ export default function KnowMore() {
         />
       </Head>
 
-      <main>
+      <>
         <section className="bg-gray-50 py-10 pt-16">
           <Hero />
         </section>
@@ -65,7 +65,7 @@ export default function KnowMore() {
         <section className="bg-white py-10 ">
           <Publications />
         </section>
-      </main>
-    </div>
+      </>
+    </>
   );
 }